Moved meowhash to separate file since it requires eas and sse4.1
