+++ lib/longtail_lib.c
@@ -5,26 +5,13 @@
 #include "../src/longtail.h"
 #include "../src/stb_ds.h"
 
-#if defined(__GNUC__) || defined(__clang__)
-    #if defined(__SSE4_1__) && defined(__AES__)
-        #define CAN_COMPILE_MEOWHASH
-    #endif
-#else
-    #define CAN_COMPILE_MEOWHASH
-#endif
-
 #include "longtail_platform.h"
 
 #include "../third-party/bikeshed/bikeshed.h"
 #include "../third-party/lizard/lib/lizard_common.h"
 #include "../third-party/lizard/lib/lizard_decompress.h"
 #include "../third-party/lizard/lib/lizard_compress.h"
-
 #include "../third-party/lizard/lib/xxhash/xxhash.h"
-
-#if defined(CAN_COMPILE_MEOWHASH)
-    #include "../third-party/meow_hash/meow_hash_x64_aesni.h"
-#endif
 
 #include <stdio.h>
 #include <errno.h>
@@ -128,73 +115,9 @@ static void ThreadWorker_Dispose(struct ThreadWorker* thread_worker)
     ThreadWorker_DisposeThread(thread_worker);
 }
 
-struct ManagedHashAPI
-{
-    struct Longtail_HashAPI m_API;
-    void (*Dispose)(struct ManagedHashAPI* managed_hash_api);
-};
-
-#if defined(CAN_COMPILE_MEOWHASH)
-
-struct MeowHashAPI
-{
-    struct ManagedHashAPI m_ManagedAPI;
-};
-
-static int MeowHash_BeginContext(struct Longtail_HashAPI* hash_api, Longtail_HashAPI_HContext* out_context)
-{
-    meow_state* state = (meow_state*)Longtail_Alloc(sizeof(meow_state));
-    MeowBegin(state, MeowDefaultSeed);
-    *out_context = (Longtail_HashAPI_HContext)state;
-    return 0;
-}
-
-static void MeowHash_Hash(struct Longtail_HashAPI* hash_api, Longtail_HashAPI_HContext context, uint32_t length, void* data)
-{
-    meow_state* state = (meow_state*)context;
-    MeowAbsorb(state, length, data);
-}
-
-static uint64_t MeowHash_EndContext(struct Longtail_HashAPI* hash_api, Longtail_HashAPI_HContext context)
-{
-    meow_state* state = (meow_state*)context;
-    uint64_t hash = (uint64_t)MeowU64From(MeowEnd(state, 0), 0);
-	Longtail_Free(state);
-    return hash;
-}
-
-static int MeowHash_HashBuffer(struct Longtail_HashAPI* hash_api, uint32_t length, void* data, uint64_t* out_hash)
-{
-    meow_state state;
-    MeowBegin(&state, MeowDefaultSeed);
-    MeowAbsorb(&state, length, data);
-    *out_hash = MeowU64From(MeowEnd(&state, 0), 0);
-    return 0;
-}
-
-static void MeowHash_Dispose(struct ManagedHashAPI* hash_api)
-{
-}
-
-static void MeowHash_Init(struct MeowHashAPI* hash_api)
-{
-    hash_api->m_ManagedAPI.m_API.BeginContext = MeowHash_BeginContext;
-    hash_api->m_ManagedAPI.m_API.Hash = MeowHash_Hash;
-    hash_api->m_ManagedAPI.m_API.EndContext = MeowHash_EndContext;
-    hash_api->m_ManagedAPI.m_API.HashBuffer = MeowHash_HashBuffer;
-    hash_api->m_ManagedAPI.Dispose = MeowHash_Dispose;
-}
-#endif // defined(CAN_COMPILE_MEOWHASH)
-
-
-
-////////////////////////////////////
-
-
-
 struct XXHashAPI
 {
-    struct ManagedHashAPI m_ManagedAPI;
+    struct Longtail_ManagedHashAPI m_ManagedAPI;
 };
 
 static int XXHash_BeginContext(struct Longtail_HashAPI* hash_api, Longtail_HashAPI_HContext* out_context)
@@ -230,7 +153,7 @@ static int XXHash_HashBuffer(struct Longtail_HashAPI* hash_api, uint32_t length,
     return 0;
 }
 
-static void XXHash_Dispose(struct ManagedHashAPI* hash_api)
+static void XXHash_Dispose(struct Longtail_ManagedHashAPI* hash_api)
 {
 }
 
@@ -243,22 +166,6 @@ static void XXHash_Init(struct XXHashAPI* hash_api)
     hash_api->m_ManagedAPI.Dispose = XXHash_Dispose;
 }
 
-
-
-////////////////////////////////////
-
-
-struct Longtail_HashAPI* Longtail_CreateMeowHashAPI()
-{
-#if defined(CAN_COMPILE_MEOWHASH)
-    struct MeowHashAPI* meow_hash = (struct MeowHashAPI*)Longtail_Alloc(sizeof(struct MeowHashAPI));
-    MeowHash_Init(meow_hash);
-    return &meow_hash->m_ManagedAPI.m_API;
-#else
-    return 0;
-#endif
-}
-
 struct Longtail_HashAPI* Longtail_CreateXXHashAPI()
 {
     struct XXHashAPI* xx_hash = (struct XXHashAPI*)Longtail_Alloc(sizeof(struct XXHashAPI));
@@ -268,7 +175,7 @@ struct Longtail_HashAPI* Longtail_CreateXXHashAPI()
 
 void Longtail_DestroyHashAPI(struct Longtail_HashAPI* hash_api)
 {
-    struct ManagedHashAPI* managed = (struct ManagedHashAPI*)hash_api;
+    struct Longtail_ManagedHashAPI* managed = (struct Longtail_ManagedHashAPI*)hash_api;
     managed->Dispose(managed);
     Longtail_Free(hash_api);
 }
@@ -282,18 +189,12 @@ void Longtail_DestroyHashAPI(struct Longtail_HashAPI* hash_api)
 
 
 
-struct ManagedStorageAPI
-{
-    struct Longtail_StorageAPI m_API;
-    void (*Dispose)(struct ManagedStorageAPI* managed_storage_api);
-};
-
 struct FSStorageAPI
 {
-    struct ManagedStorageAPI m_StorageAPI;
+    struct Longtail_ManagedStorageAPI m_StorageAPI;
 };
 
-static void FSStorageAPI_Dispose(struct ManagedStorageAPI* storage_api)
+static void FSStorageAPI_Dispose(struct Longtail_ManagedStorageAPI* storage_api)
 {
 }
 
@@ -508,14 +409,14 @@ struct Lookup
 
 struct InMemStorageAPI
 {
-    struct ManagedStorageAPI m_StorageAPI;
+    struct Longtail_ManagedStorageAPI m_StorageAPI;
     struct Longtail_HashAPI* m_HashAPI;
     struct Lookup* m_PathHashToContent;
     struct PathEntry* m_PathEntries;
     HLongtail_SpinLock m_SpinLock;
 };
 
-static void InMemStorageAPI_Dispose(struct ManagedStorageAPI* storage_api)
+static void InMemStorageAPI_Dispose(struct Longtail_ManagedStorageAPI* storage_api)
 {
     struct InMemStorageAPI* in_mem_storage_api = (struct InMemStorageAPI*)storage_api;
     size_t c = (size_t)arrlen(in_mem_storage_api->m_PathEntries);
@@ -979,7 +880,7 @@ static void InMemStorageAPI_Init(struct InMemStorageAPI* storage_api)
     storage_api->m_StorageAPI.m_API.GetEntrySize = InMemStorageAPI_GetEntrySize;
     storage_api->m_StorageAPI.Dispose = InMemStorageAPI_Dispose;
 
-    storage_api->m_HashAPI = Longtail_CreateMeowHashAPI();
+    storage_api->m_HashAPI = Longtail_CreateXXHashAPI();
     storage_api->m_PathHashToContent = 0;
     storage_api->m_PathEntries = 0;
     int err = Longtail_CreateSpinLock(&storage_api[1], &storage_api->m_SpinLock);
@@ -1000,7 +901,7 @@ struct Longtail_StorageAPI* Longtail_CreateInMemStorageAPI()
 
 void Longtail_DestroyStorageAPI(struct Longtail_StorageAPI* storage_api)
 {
-    struct ManagedStorageAPI* managed = (struct ManagedStorageAPI*)storage_api;
+    struct Longtail_ManagedStorageAPI* managed = (struct Longtail_ManagedStorageAPI*)storage_api;
     managed->Dispose(managed);
     Longtail_Free(managed);
 }
@@ -1008,12 +909,6 @@ void Longtail_DestroyStorageAPI(struct Longtail_StorageAPI* storage_api)
 
 
 
-
-struct ManagedJobAPI
-{
-    struct Longtail_JobAPI m_API;
-    void (*Dispose)(struct ManagedJobAPI* managed_job_api);
-};
 
 struct JobWrapper
 {
@@ -1024,7 +919,7 @@ struct JobWrapper
 
 struct BikeshedJobAPI
 {
-    struct ManagedJobAPI m_ManagedAPI;
+    struct Longtail_ManagedJobAPI m_ManagedAPI;
 
     struct ReadyCallback m_ReadyCallback;
     Bikeshed m_Shed;
@@ -1178,7 +1073,7 @@ static int Bikeshed_WaitForAllJobs(struct Longtail_JobAPI* job_api, void* contex
     return 0;
 }
 
-static void Bikeshed_Dispose(struct ManagedJobAPI* job_api)
+static void Bikeshed_Dispose(struct Longtail_ManagedJobAPI* job_api)
 {
     struct BikeshedJobAPI* bikeshed_job_api = (struct BikeshedJobAPI*)job_api;
     Longtail_AtomicAdd32(&bikeshed_job_api->m_Stop, 1);
@@ -1236,23 +1131,17 @@ struct Longtail_JobAPI* Longtail_CreateBikeshedJobAPI(uint32_t worker_count)
 
 void Longtail_DestroyJobAPI(struct Longtail_JobAPI* job_api)
 {
-    struct ManagedJobAPI* managed = (struct ManagedJobAPI*)job_api;
+    struct Longtail_ManagedJobAPI* managed = (struct Longtail_ManagedJobAPI*)job_api;
     managed->Dispose(managed);
     Longtail_Free(job_api);
 }
 
-struct ManagedCompressionAPI
-{
-    struct Longtail_CompressionAPI m_API;
-    void (*Dispose)(struct ManagedCompressionAPI* managed_compression_api);
-};
-
 struct LizardCompressionAPI
 {
-    struct ManagedCompressionAPI m_CompressionAPI;
+    struct Longtail_ManagedCompressionAPI m_CompressionAPI;
 };
 
-void LizardCompressionAPI_Dispose(struct ManagedCompressionAPI* compression_api)
+void LizardCompressionAPI_Dispose(struct Longtail_ManagedCompressionAPI* compression_api)
 {
 }
 
@@ -1339,7 +1228,7 @@ struct Longtail_CompressionAPI* Longtail_CreateLizardCompressionAPI()
 
 void Longtail_DestroyCompressionAPI(struct Longtail_CompressionAPI* compression_api)
 {
-    struct ManagedCompressionAPI* managed = (struct ManagedCompressionAPI*)compression_api;
+    struct Longtail_ManagedCompressionAPI* managed = (struct Longtail_ManagedCompressionAPI*)compression_api;
     managed->Dispose(managed);
     Longtail_Free(managed);
 }
