+++ lib/longtail_meowhash.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+extern struct Longtail_HashAPI* Longtail_CreateMeowHashAPI();
+
+#ifdef __cplusplus
+}
+#endif
